refactor(im-api): 用户/消息/会话相关 ID 统一为 Long，修复大 ID 用户 400
chat_user 用雪花 ID，AUTO_INCREMENT 已达 2091818653023719424（远超
Integer.MAX_VALUE）。凡是 ID 声明为 Integer 的位置，大 ID 用户都会失败：

- ChatSessionMapper：getSessionsByUserId / getOrCreateSession /
  updateLastMessage / clearUnreadCount 的 userId、targetId 改 Long
- ChatMessageMapper：getHistoryMessages 的 userId、targetId 改 Long
- ChatMessageService(+Impl)：getHistoryMessages / getHistoryMessagesByCursor
  的 userId、targetId、beforeMessageId 改 Long
- ChatMessageController：history / historyCursor 的 @RequestParam
  userId、targetId、beforeMessageId 改 Long（page/size 仍是 Integer）
- ChatUserVO.id 改 Long（返回给前端的用户 ID，Integer 会溢出）

数据库侧：V2 建表时 chat_group.owner_id 误建为 int，与实体
ChatGroup.ownerId(Long) 不匹配，大 ID 用户建群会失败。新增 V4 迁移改为
bigint（表为空、无外键，可安全 MODIFY）。其余存 ID 的列已是 bigint，已全表核对。

补 MapperIdTypeTest：用反射钉住上述方法/实体的 ID 类型，并断言 -parameters
编译参数可用（避免参数名退化成 arg0 导致测试假通过）；另加雪花 ID 透传与
游标分页的控制器用例。

+++ snow-chat-im-backend/snow-im-api/src/main/java/com/stylesmile/chat/controller/ChatMessageController.java
@@ -47,8 +47,8 @@ public class ChatMessageController {
     @ApiResponse(responseCode = "200", description = "查询成功")
     @GetMapping("/history")
     public Result<List<ChatMessage>> history(
-            @Parameter(description = "用户ID", required = true) @RequestParam Integer userId,
-            @Parameter(description = "目标ID（好友ID或群组ID）", required = true) @RequestParam Integer targetId,
+            @Parameter(description = "用户ID", required = true) @RequestParam Long userId,
+            @Parameter(description = "目标ID（好友ID或群组ID）", required = true) @RequestParam Long targetId,
             @Parameter(description = "目标类型：friend/group/file_helper", required = true) @RequestParam String targetType,
             @Parameter(description = "页码，从1开始", example = "1") @RequestParam(defaultValue = "1") Integer page,
             @Parameter(description = "每页条数", example = "20") @RequestParam(defaultValue = "20") Integer size) {
@@ -63,10 +63,10 @@ public class ChatMessageController {
     @ApiResponse(responseCode = "200", description = "查询成功")
     @GetMapping("/history/cursor")
     public Result<List<ChatMessage>> historyCursor(
-            @Parameter(description = "用户ID", required = true) @RequestParam Integer userId,
-            @Parameter(description = "目标ID", required = true) @RequestParam Integer targetId,
+            @Parameter(description = "用户ID", required = true) @RequestParam Long userId,
+            @Parameter(description = "目标ID", required = true) @RequestParam Long targetId,
             @Parameter(description = "目标类型：friend/group/file_helper", required = true) @RequestParam String targetType,
-            @Parameter(description = "上一页最后一条消息ID，用于游标分页") @RequestParam(required = false) Integer beforeMessageId,
+            @Parameter(description = "上一页最后一条消息ID，用于游标分页") @RequestParam(required = false) Long beforeMessageId,
             @Parameter(description = "每页条数", example = "20") @RequestParam(defaultValue = "20") Integer size) {
         List<ChatMessage> messages = chatMessageService.getHistoryMessagesByCursor(userId, targetId, targetType, beforeMessageId, size);
         return Result.success(messages);
